# server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RentGraph API",
    "version": "1.1.0"
  },
  "paths": {
...
    "/api/v1/houses": {
      "get": {
        "tags": [
          "houses"
        ],
        "summary": "List Houses",
        "operationId": "list_houses_api_v1_houses_get",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "active",
                "dropped",
                "all"
              ],
              "type": "string",
              "default": "active",
              "title": "Status"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HouseOut"
                  },
                  "title": "Response List Houses Api V1 Houses Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "houses"
        ],
        "summary": "Create House",
        "operationId": "create_house_api_v1_houses_post",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HouseCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HouseOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/houses/{house_id}": {
      "get": {
        "tags": [
          "houses"
        ],
        "summary": "Get House",
        "operationId": "get_house_api_v1_houses__house_id__get",
        "parameters": [
          {
            "name": "house_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "House Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HouseOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "houses"
        ],
        "summary": "Patch House",
        "operationId": "patch_house_api_v1_houses__house_id__patch",
        "parameters": [
          {
            "name": "house_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "House Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HousePatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HouseOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "houses"
        ],
        "summary": "Delete House",
        "operationId": "delete_house_api_v1_houses__house_id__delete",
        "parameters": [
          {
            "name": "house_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "House Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/houses/{house_id}/evidence": {
      "post": {
        "tags": [
          "houses"
        ],
        "summary": "Add Evidence",
        "description": "补充字段证据（用户手动修正字段后可记录来源，保证可追溯）。",
        "operationId": "add_evidence_api_v1_houses__house_id__evidence_post",
        "parameters": [
          {
            "name": "house_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "House Id"
            }
          },
          {
            "name": "field",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Field"
            }
          },
          {
            "name": "snippet",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Snippet"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HouseOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/preferences": {
      "get": {
        "tags": [
          "preferences"
        ],
        "summary": "Get Preferences",
        "operationId": "get_preferences_api_v1_preferences_get",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferenceOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "preferences"
        ],
        "summary": "Put Preferences",
        "operationId": "put_preferences_api_v1_preferences_put",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreferenceIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferenceOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HouseCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "default": ""
          },
          "region": {
            "type": "string",
            "title": "Region",
            "default": ""
          },
          "address": {
            "type": "string",
            "title": "Address",
            "default": ""
          },
          "rent": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rent"
          },
          "deposit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deposit"
          },
          "agency_fee": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agency Fee"
          },
          "property_fee": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Property Fee"
          },
          "property_bear": {
            "type": "string",
            "title": "Property Bear",
            "default": ""
          },
          "net_fee": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Net Fee"
          },
          "other_fee": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Other Fee"
          },
          "area": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Area"
          },
          "layout": {
            "type": "string",
            "title": "Layout",
            "default": ""
          },
          "floor": {
            "type": "string",
            "title": "Floor",
            "default": ""
          },
          "orientation": {
            "type": "string",
            "title": "Orientation",
            "default": ""
          },
          "bathroom": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bathroom"
          },
          "lighting": {
            "type": "string",
            "title": "Lighting",
            "default": ""
          },
          "furniture": {
            "type": "string",
            "title": "Furniture",
            "default": ""
          },
          "commute_min": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Commute Min"
          },
          "commute_mode": {
            "type": "string",
            "title": "Commute Mode",
            "default": "地铁"
          },
          "metro": {
            "type": "string",
            "title": "Metro",
            "default": ""
          },
          "nearby": {
            "type": "string",
            "title": "Nearby",
            "default": ""
          },
          "pet": {
            "type": "string",
            "title": "Pet",
            "default": ""
          },
          "shared": {
            "type": "boolean",
            "title": "Shared",
            "default": false
          },
          "sublet": {
            "type": "string",
            "title": "Sublet",
            "default": ""
          },
          "max_people": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max People"
          },
          "lease_req": {
            "type": "string",
            "title": "Lease Req",
            "default": ""
          },
          "available": {
            "type": "string",
            "title": "Available",
            "default": ""
          },
          "notes": {
            "type": "string",
            "title": "Notes",
            "default": ""
          },
          "link_url": {
            "type": "string",
            "title": "Link Url",
            "default": ""
          },
          "source": {
            "type": "string",
            "enum": [
              "paste",
              "manual",
              "batch",
              "file-xlsx",
              "file-docx",
              "file-txt",
              "link"
            ],
            "title": "Source",
            "default": "manual"
          },
          "raw": {
            "type": "string",
            "title": "Raw",
            "default": ""
          }
        },
        "type": "object",
        "title": "HouseCreate"
      },
...
      "HouseOut": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "default": ""
          },
          "region": {
            "type": "string",
            "title": "Region",
            "default": ""
          },
          "address": {
            "type": "string",
            "title": "Address",
            "default": ""
          },
          "rent": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rent"
          },
          "deposit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deposit"
          },
          "agency_fee": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agency Fee"
          },
          "property_fee": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Property Fee"
          },
          "property_bear": {
            "type": "string",
            "title": "Property Bear",
            "default": ""
          },
          "net_fee": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Net Fee"
          },
          "other_fee": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Other Fee"
          },
          "area": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Area"
          },
          "layout": {
            "type": "string",
            "title": "Layout",
            "default": ""
          },
          "floor": {
            "type": "string",
            "title": "Floor",
            "default": ""
          },
          "orientation": {
            "type": "string",
            "title": "Orientation",
            "default": ""
          },
          "bathroom": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bathroom"
          },
          "lighting": {
            "type": "string",
            "title": "Lighting",
            "default": ""
          },
          "furniture": {
            "type": "string",
            "title": "Furniture",
            "default": ""
          },
          "commute_min": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Commute Min"
          },
          "commute_mode": {
            "type": "string",
            "title": "Commute Mode",
            "default": "地铁"
          },
          "metro": {
            "type": "string",
            "title": "Metro",
            "default": ""
          },
          "nearby": {
            "type": "string",
            "title": "Nearby",
            "default": ""
          },
          "pet": {
            "type": "string",
            "title": "Pet",
            "default": ""
          },
          "shared": {
            "type": "boolean",
            "title": "Shared",
            "default": false
          },
          "sublet": {
            "type": "string",
            "title": "Sublet",
            "default": ""
          },
          "max_people": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max People"
          },
          "lease_req": {
            "type": "string",
            "title": "Lease Req",
            "default": ""
          },
          "available": {
            "type": "string",
            "title": "Available",
            "default": ""
          },
          "notes": {
            "type": "string",
            "title": "Notes",
            "default": ""
          },
          "link_url": {
            "type": "string",
            "title": "Link Url",
            "default": ""
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "no": {
            "type": "string",
            "title": "No"
          },
          "batch_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Batch Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "dropped"
            ],
            "title": "Status",
            "default": "active"
          },
          "drop_reason": {
            "type": "string",
            "title": "Drop Reason",
            "default": ""
          },
          "source": {
            "type": "string",
            "enum": [
              "paste",
              "manual",
              "batch",
              "file-xlsx",
              "file-docx",
              "file-txt",
              "link"
            ],
            "title": "Source",
            "default": "paste"
          },
          "raw": {
            "type": "string",
            "title": "Raw",
            "default": ""
          },
          "visits": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Visits"
          },
          "todos": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Todos"
          },
          "verify": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Verify"
          },
          "contract_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contract Id"
          },
          "source_label": {
            "type": "string",
            "title": "Source Label",
            "default": ""
          },
          "completeness": {
            "type": "integer",
            "title": "Completeness",
            "default": 0
          },
          "missing": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing"
          },
          "evidence": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Evidence"
          }
        },
        "type": "object",
        "required": [
          "id",
          "no"
        ],
        "title": "HouseOut"
      },
...
      "PreferenceIn": {
        "properties": {
          "budget": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 1000000.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget",
            "description": "硬约束：月租上限（元）"
          },
          "commute": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 600.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Commute",
            "description": "硬约束：通勤上限（分钟）"
          },
          "need_bathroom": {
            "type": "boolean",
            "title": "Need Bathroom",
            "default": false
          },
          "allow_shared": {
            "type": "boolean",
            "title": "Allow Shared",
            "default": true
          },
          "soft": {
            "$ref": "#/components/schemas/SoftPrefs"
          }
        },
        "type": "object",
        "title": "PreferenceIn"
      },
      "PreferenceOut": {
        "properties": {
          "budget": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 1000000.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget",
            "description": "硬约束：月租上限（元）"
          },
          "commute": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 600.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Commute",
            "description": "硬约束：通勤上限（分钟）"
          },
          "need_bathroom": {
            "type": "boolean",
            "title": "Need Bathroom",
            "default": false
          },
          "allow_shared": {
            "type": "boolean",
            "title": "Allow Shared",
            "default": true
          },
          "soft": {
            "$ref": "#/components/schemas/SoftPrefs"
          },
          "version": {
            "type": "integer",
            "title": "Version",
            "default": 1
          }
        },
        "type": "object",
        "title": "PreferenceOut"
      },
...
      "SoftPrefs": {
        "properties": {
          "south": {
            "type": "boolean",
            "title": "South",
            "default": false
          },
          "light": {
            "type": "boolean",
            "title": "Light",
            "default": false
          },
          "high_floor": {
            "type": "boolean",
            "title": "High Floor",
            "default": false
          },
          "big_area": {
            "type": "boolean",
            "title": "Big Area",
            "default": false
          },
          "flex_pay": {
            "type": "boolean",
            "title": "Flex Pay",
            "default": false
          }
        },
        "type": "object",
        "title": "SoftPrefs"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}